test: get_forks
