+++ chain/src/chain.rs
@@ -51,10 +51,32 @@ pub struct BlockInsertionResult {
     pub new_best_block: bool,
 }
 
-struct Fork {
-    new_blocks: Vec<Block>,
-    old_blocks: Vec<Block>,
-    open_exts: Vec<BlockExt>,
+#[derive(Debug, Default)]
+pub(crate) struct Fork {
+    pub(crate) new_blocks: Vec<Block>,
+    pub(crate) old_blocks: Vec<Block>,
+    pub(crate) open_exts: Vec<BlockExt>,
+}
+
+pub(crate) struct GlobalIndex {
+    pub(crate) number: BlockNumber,
+    pub(crate) hash: H256,
+    pub(crate) unseen: bool,
+}
+
+impl GlobalIndex {
+    pub(crate) fn new(number: BlockNumber, hash: H256, unseen: bool) -> GlobalIndex {
+        GlobalIndex {
+            number,
+            hash,
+            unseen,
+        }
+    }
+
+    pub(crate) fn forward(&mut self, hash: H256) {
+        self.number -= 1;
+        self.hash = hash;
+    }
 }
 
 pub struct ChainService<CI> {
@@ -119,7 +141,7 @@ impl<CI: ChainIndex + 'static> ChainService<CI> {
 
     // process_block will do block verify
     // but invoker should guarantee block header be verified
-    fn process_block(&mut self, block: Arc<Block>) -> Result<(), ProcessBlockError> {
+    pub(crate) fn process_block(&mut self, block: Arc<Block>) -> Result<(), ProcessBlockError> {
         debug!(target: "chain", "begin processing block: {}", block.header().hash());
         if self.verification {
             let block_verifier = BlockVerifier::new(self.shared.clone());
@@ -136,7 +158,7 @@ impl<CI: ChainIndex + 'static> ChainService<CI> {
     }
 
     #[allow(clippy::op_ref)]
-    fn insert_block(&self, block: &Block) -> Result<BlockInsertionResult, SharedError> {
+    pub(crate) fn insert_block(&self, block: &Block) -> Result<BlockInsertionResult, SharedError> {
         let mut new_best_block = false;
         let mut total_difficulty = U256::zero();
 
@@ -227,7 +249,7 @@ impl<CI: ChainIndex + 'static> ChainService<CI> {
         })
     }
 
-    fn post_insert_result(&mut self, block: Arc<Block>, result: BlockInsertionResult) {
+    pub(crate) fn post_insert_result(&mut self, block: Arc<Block>, result: BlockInsertionResult) {
         let BlockInsertionResult {
             new_best_block,
             mut fork_blks,
@@ -247,7 +269,12 @@ impl<CI: ChainIndex + 'static> ChainService<CI> {
         }
     }
 
-    fn update_index(&self, batch: &mut Batch, old_blocks: &[Block], new_blocks: &[Block]) {
+    pub(crate) fn update_index(
+        &self,
+        batch: &mut Batch,
+        old_blocks: &[Block],
+        new_blocks: &[Block],
+    ) {
         let old_number = match old_blocks.get(0) {
             Some(b) => b.header().number(),
             None => 0,
@@ -283,97 +310,103 @@ impl<CI: ChainIndex + 'static> ChainService<CI> {
         }
     }
 
-    fn get_forks(
+    fn alignment_fork(
+        &self,
+        fork: &mut Fork,
+        index: &mut GlobalIndex,
+        new_tip_number: BlockNumber,
+        current_tip_number: BlockNumber,
+    ) {
+        if new_tip_number <= current_tip_number {
+            for bn in new_tip_number..=current_tip_number {
+                let hash = self.shared.block_hash(bn).unwrap();
+                let old_block = self.shared.block(&hash).unwrap();
+                fork.old_blocks.push(old_block);
+            }
+        } else {
+            while index.number > current_tip_number {
+                if index.unseen {
+                    let ext = self.shared.block_ext(&index.hash).unwrap();
+                    if ext.valid.is_none() {
+                        fork.open_exts.push(ext)
+                    } else {
+                        index.unseen = false;
+                    }
+                }
+                let new_block = self.shared.block(&index.hash).unwrap();
+                index.forward(new_block.header().parent_hash().clone());
+                fork.new_blocks.push(new_block);
+            }
+        }
+    }
+
+    fn find_fork_until_latest_common(&self, fork: &mut Fork, index: &mut GlobalIndex) {
+        loop {
+            if index.number == 0 {
+                break;
+            }
+            let old_hash = self.shared.block_hash(index.number).unwrap();
+            if old_hash == index.hash {
+                break;
+            }
+            let old_block = self.shared.block(&old_hash).unwrap();
+            fork.old_blocks.push(old_block);
+
+            if index.unseen {
+                let ext = self.shared.block_ext(&index.hash).unwrap();
+                if ext.valid.is_none() {
+                    fork.open_exts.push(ext)
+                } else {
+                    index.unseen = false;
+                }
+            }
+
+            let new_block = self.shared.block(&index.hash).unwrap();
+            index.forward(new_block.header().parent_hash().clone());
+            fork.new_blocks.push(new_block);
+        }
+    }
+
+    pub(crate) fn find_fork(
         &self,
         current_tip_number: BlockNumber,
         new_tip_block: &Block,
         new_tip_ext: BlockExt,
     ) -> Option<Fork> {
         let new_tip_number = new_tip_block.header().number();
-        let mut old_blocks = Vec::new();
-        let mut new_blocks = Vec::new();
-        let mut open_exts = Vec::new();
+        let mut fork = Fork::default();
 
         if new_tip_ext.valid.is_none() {
-            open_exts.push(new_tip_ext);
+            fork.open_exts.push(new_tip_ext);
         } else {
             // txs in block are invalid
             return None;
         }
 
-        let mut is_new_record = true;
-
         // new_blocks = forks[latest_common + 1 .. new_tip]
         // old_blocks = chain[latest_common + 1 .. old_tip]
-        new_blocks.push(new_tip_block.clone());
-        let mut index = new_tip_number - 1;
-        let mut index_hash = new_tip_block.header().parent_hash().clone();
+        fork.new_blocks.push(new_tip_block.clone());
+
+        let mut index = GlobalIndex::new(
+            new_tip_number - 1,
+            new_tip_block.header().parent_hash().clone(),
+            true,
+        );
 
         // if new_tip_number <= current_tip_number
         // then old_blocks.extend(chain[new_tip_number .. =current_tip_number])
         // if new_tip_number > current_tip_number
         // then new_blocks.extend(forks[current_tip_number + 1 .. =new_tip_number])
-        if new_tip_number <= current_tip_number {
-            for bn in new_tip_number..=current_tip_number {
-                let hash = self.shared.block_hash(bn).unwrap();
-                let old_block = self.shared.block(&hash).unwrap();
-                old_blocks.push(old_block);
-            }
-        } else {
-            while index > current_tip_number {
-                let new_block = self.shared.block(&index_hash).unwrap();
-
-                if is_new_record {
-                    let ext = self.shared.block_ext(&index_hash).unwrap();
-                    if ext.valid.is_none() {
-                        open_exts.push(ext)
-                    } else {
-                        is_new_record = false;
-                    }
-                }
-
-                index_hash = new_block.header().parent_hash().clone();
-                new_blocks.push(new_block);
-                index -= 1;
-            }
-        }
+        self.alignment_fork(&mut fork, &mut index, new_tip_number, current_tip_number);
 
         // find latest common ancestor
-        loop {
-            if index == 0 {
-                break;
-            }
-            let old_hash = self.shared.block_hash(index).unwrap();
-            if old_hash == index_hash {
-                break;
-            }
-            let old_block = self.shared.block(&old_hash).unwrap();
-            old_blocks.push(old_block);
+        self.find_fork_until_latest_common(&mut fork, &mut index);
 
-            let new_block = self.shared.block(&index_hash).unwrap();
-            index_hash = new_block.header().parent_hash().clone();
-            new_blocks.push(new_block);
-
-            if is_new_record {
-                let ext = self.shared.block_ext(&index_hash).unwrap();
-                if ext.valid.is_none() {
-                    open_exts.push(ext)
-                } else {
-                    is_new_record = false;
-                }
-            }
-            index -= 1;
-        }
-
-        Some(Fork {
-            new_blocks,
-            old_blocks,
-            open_exts,
-        })
+        Some(fork)
     }
 
     // we found new best_block total_difficulty > old_chain.total_difficulty
-    fn reconcile_main_chain(
+    pub(crate) fn reconcile_main_chain(
         &self,
         batch: &mut Batch,
         tip_number: BlockNumber,
@@ -384,7 +417,7 @@ impl<CI: ChainIndex + 'static> ChainService<CI> {
         let skip_verify = !self.verification;
 
         let mut fork = self
-            .get_forks(tip_number, block, ext)
+            .find_fork(tip_number, block, ext)
             .ok_or(SharedError::InvalidTransaction)?;
 
         let mut old_inputs = FnvHashSet::default();
@@ -590,431 +623,5 @@ impl<CI: ChainIndex + 'static> ChainBuilder<CI> {
 
     pub fn build(self) -> ChainService<CI> {
         ChainService::new(self.shared, self.notify, self.verification)
-    }
-}
-
-#[cfg(test)]
-pub mod test {
-    use super::*;
-    use ckb_chain_spec::consensus::Consensus;
-    use ckb_core::block::BlockBuilder;
-    use ckb_core::cell::CellProvider;
-    use ckb_core::header::{Header, HeaderBuilder};
-    use ckb_core::transaction::{
-        CellInput, CellOutput, OutPoint, ProposalShortId, Transaction, TransactionBuilder,
-    };
-    use ckb_core::uncle::UncleBlock;
-    use ckb_db::memorydb::MemoryKeyValueDB;
-    use ckb_notify::NotifyService;
-    use ckb_shared::shared::SharedBuilder;
-    use ckb_shared::store::ChainKVStore;
-    use numext_fixed_uint::U256;
-
-    fn start_chain(
-        consensus: Option<Consensus>,
-    ) -> (ChainController, Shared<ChainKVStore<MemoryKeyValueDB>>) {
-        let builder = SharedBuilder::<MemoryKeyValueDB>::new();
-        let shared = builder
-            .consensus(consensus.unwrap_or_else(Default::default))
-            .build();
-
-        let notify = NotifyService::default().start::<&str>(None);
-        let chain_service = ChainBuilder::new(shared.clone(), notify)
-            .verification(false)
-            .build();
-        let chain_controller = chain_service.start::<&str>(None);
-        (chain_controller, shared)
-    }
-
-    fn create_cellbase(number: BlockNumber) -> Transaction {
-        TransactionBuilder::default()
-            .input(CellInput::new_cellbase_input(number))
-            .output(CellOutput::new(0, vec![], H256::zero(), None))
-            .build()
-    }
-
-    fn gen_block(
-        parent_header: &Header,
-        nonce: u64,
-        difficulty: U256,
-        commit_transactions: Vec<Transaction>,
-        uncles: Vec<UncleBlock>,
-    ) -> Block {
-        let number = parent_header.number() + 1;
-        let cellbase = create_cellbase(number);
-        let header = HeaderBuilder::default()
-            .parent_hash(parent_header.hash().clone())
-            .timestamp(unix_time_as_millis())
-            .number(number)
-            .difficulty(difficulty)
-            .nonce(nonce)
-            .build();
-
-        BlockBuilder::default()
-            .header(header)
-            .commit_transaction(cellbase)
-            .commit_transactions(commit_transactions)
-            .uncles(uncles)
-            .proposal_transaction(ProposalShortId::from_slice(&[1; 10]).unwrap())
-            .build()
-    }
-
-    fn create_transaction(parent: H256) -> Transaction {
-        let mut output = CellOutput::default();
-        output.capacity = 100_000_000 / 100 as u64;
-        let outputs: Vec<CellOutput> = vec![output.clone(); 100];
-
-        TransactionBuilder::default()
-            .input(CellInput::new(OutPoint::new(parent, 0), Default::default()))
-            .outputs(outputs)
-            .build()
-    }
-
-    #[test]
-    fn test_genesis_transaction_spend() {
-        let tx = TransactionBuilder::default()
-            .input(CellInput::new(OutPoint::null(), Default::default()))
-            .outputs(vec![
-                CellOutput::new(
-                    100_000_000,
-                    vec![],
-                    H256::default(),
-                    None
-                );
-                100
-            ])
-            .build();
-
-        let mut root_hash = tx.hash().clone();
-
-        let genesis_block = BlockBuilder::default()
-            .commit_transaction(tx)
-            .with_header_builder(HeaderBuilder::default().difficulty(U256::from(1000u64)));
-
-        let consensus = Consensus::default().set_genesis_block(genesis_block);
-        let (chain_controller, shared) = start_chain(Some(consensus));
-
-        let end = 21;
-
-        let mut blocks1: Vec<Block> = vec![];
-        let mut parent = shared.block_header(&shared.block_hash(0).unwrap()).unwrap();
-        for i in 1..end {
-            let difficulty = parent.difficulty().clone();
-            let tx = create_transaction(root_hash);
-            root_hash = tx.hash().clone();
-            let new_block = gen_block(&parent, i, difficulty + U256::from(1u64), vec![tx], vec![]);
-            blocks1.push(new_block.clone());
-            parent = new_block.header().clone();
-        }
-
-        for block in &blocks1[0..10] {
-            assert!(chain_controller
-                .process_block(Arc::new(block.clone()))
-                .is_ok());
-        }
-    }
-
-    #[test]
-    fn test_genesis_transaction_fetch() {
-        let tx = TransactionBuilder::default()
-            .input(CellInput::new(OutPoint::null(), Default::default()))
-            .outputs(vec![
-                CellOutput::new(
-                    100_000_000,
-                    vec![],
-                    H256::default(),
-                    None
-                );
-                100
-            ])
-            .build();
-
-        let root_hash = tx.hash().clone();
-
-        let genesis_block = BlockBuilder::default()
-            .commit_transaction(tx)
-            .with_header_builder(HeaderBuilder::default().difficulty(U256::from(1000u64)));
-
-        let consensus = Consensus::default().set_genesis_block(genesis_block);
-        let (_chain_controller, shared) = start_chain(Some(consensus));
-
-        let out_point = OutPoint::new(root_hash, 0);
-        let state = shared.cell(&out_point);
-        assert!(state.is_live());
-    }
-
-    #[test]
-    fn test_chain_fork_by_total_difficulty() {
-        let (chain_controller, shared) = start_chain(None);
-        let final_number = 20;
-
-        let mut chain1: Vec<Block> = Vec::new();
-        let mut chain2: Vec<Block> = Vec::new();
-
-        let mut parent = shared.block_header(&shared.block_hash(0).unwrap()).unwrap();
-        for i in 1..final_number {
-            let difficulty = parent.difficulty().clone();
-            let new_block = gen_block(&parent, i, difficulty + U256::from(100u64), vec![], vec![]);
-            chain1.push(new_block.clone());
-            parent = new_block.header().clone();
-        }
-
-        parent = shared.block_header(&shared.block_hash(0).unwrap()).unwrap();
-        for i in 1..final_number {
-            let difficulty = parent.difficulty().clone();
-            let j = if i > 10 { 110 } else { 99 };
-            let new_block = gen_block(
-                &parent,
-                i + 1000,
-                difficulty + U256::from(j as u32),
-                vec![],
-                vec![],
-            );
-            chain2.push(new_block.clone());
-            parent = new_block.header().clone();
-        }
-
-        for block in &chain1 {
-            chain_controller
-                .process_block(Arc::new(block.clone()))
-                .expect("process block ok");
-        }
-
-        for block in &chain2 {
-            chain_controller
-                .process_block(Arc::new(block.clone()))
-                .expect("process block ok");
-        }
-        assert_eq!(
-            shared.block_hash(8),
-            chain2.get(7).map(|b| b.header().hash())
-        );
-    }
-
-    #[test]
-    fn test_chain_fork_by_hash() {
-        let (chain_controller, shared) = start_chain(None);
-        let final_number = 20;
-
-        let mut chain1: Vec<Block> = Vec::new();
-        let mut chain2: Vec<Block> = Vec::new();
-
-        let mut parent = shared.block_header(&shared.block_hash(0).unwrap()).unwrap();
-        for i in 1..final_number {
-            let difficulty = parent.difficulty().clone();
-            let new_block = gen_block(&parent, i, difficulty + U256::from(100u64), vec![], vec![]);
-            chain1.push(new_block.clone());
-            parent = new_block.header().clone();
-        }
-
-        parent = shared.block_header(&shared.block_hash(0).unwrap()).unwrap();
-        for i in 1..final_number {
-            let difficulty = parent.difficulty().clone();
-            let new_block = gen_block(
-                &parent,
-                i + 1000,
-                difficulty + U256::from(100u64),
-                vec![],
-                vec![],
-            );
-            chain2.push(new_block.clone());
-            parent = new_block.header().clone();
-        }
-
-        for block in &chain1 {
-            chain_controller
-                .process_block(Arc::new(block.clone()))
-                .expect("process block ok");
-        }
-
-        for block in &chain2 {
-            chain_controller
-                .process_block(Arc::new(block.clone()))
-                .expect("process block ok");
-        }
-
-        //if total_difficulty equal, we chose block which have smaller hash as best
-        assert!(chain1
-            .iter()
-            .zip(chain2.iter())
-            .all(|(a, b)| a.header().difficulty() == b.header().difficulty()));
-
-        let best = if chain1[(final_number - 2) as usize].header().hash()
-            < chain2[(final_number - 2) as usize].header().hash()
-        {
-            chain1
-        } else {
-            chain2
-        };
-        assert_eq!(shared.block_hash(8), best.get(7).map(|b| b.header().hash()));
-        assert_eq!(
-            shared.block_hash(19),
-            best.get(18).map(|b| b.header().hash())
-        );
-    }
-
-    #[test]
-    fn test_chain_get_ancestor() {
-        let (chain_controller, shared) = start_chain(None);
-        let final_number = 20;
-
-        let mut chain1: Vec<Block> = Vec::new();
-        let mut chain2: Vec<Block> = Vec::new();
-
-        let mut parent = shared.block_header(&shared.block_hash(0).unwrap()).unwrap();
-        for i in 1..final_number {
-            let difficulty = parent.difficulty().clone();
-            let new_block = gen_block(&parent, i, difficulty + U256::from(100u64), vec![], vec![]);
-            chain1.push(new_block.clone());
-            parent = new_block.header().clone();
-        }
-
-        parent = shared.block_header(&shared.block_hash(0).unwrap()).unwrap();
-        for i in 1..final_number {
-            let difficulty = parent.difficulty().clone();
-            let new_block = gen_block(
-                &parent,
-                i + 1000,
-                difficulty + U256::from(100u64),
-                vec![],
-                vec![],
-            );
-            chain2.push(new_block.clone());
-            parent = new_block.header().clone();
-        }
-
-        for block in &chain1 {
-            chain_controller
-                .process_block(Arc::new(block.clone()))
-                .expect("process block ok");
-        }
-
-        for block in &chain2 {
-            chain_controller
-                .process_block(Arc::new(block.clone()))
-                .expect("process block ok");
-        }
-
-        assert_eq!(
-            *chain1[9].header(),
-            shared
-                .get_ancestor(&chain1.last().unwrap().header().hash(), 10)
-                .unwrap()
-        );
-
-        assert_eq!(
-            *chain2[9].header(),
-            shared
-                .get_ancestor(&chain2.last().unwrap().header().hash(), 10)
-                .unwrap()
-        );
-    }
-
-    #[test]
-    fn test_calculate_difficulty() {
-        let genesis_block = BlockBuilder::default()
-            .with_header_builder(HeaderBuilder::default().difficulty(U256::from(1000u64)));
-        let mut consensus = Consensus::default().set_genesis_block(genesis_block);
-        consensus.pow_time_span = 200;
-        consensus.pow_spacing = 1;
-
-        let (chain_controller, shared) = start_chain(Some(consensus.clone()));
-        let final_number = shared.consensus().difficulty_adjustment_interval();
-
-        let mut chain1: Vec<Block> = Vec::new();
-        let mut chain2: Vec<Block> = Vec::new();
-
-        let mut parent = shared.block_header(&shared.block_hash(0).unwrap()).unwrap();
-        for i in 1..final_number - 1 {
-            let difficulty = shared.calculate_difficulty(&parent).unwrap();
-            let new_block = gen_block(&parent, i, difficulty, vec![], vec![]);
-            chain_controller
-                .process_block(Arc::new(new_block.clone()))
-                .expect("process block ok");
-            chain1.push(new_block.clone());
-            parent = new_block.header().clone();
-        }
-
-        parent = shared.block_header(&shared.block_hash(0).unwrap()).unwrap();
-        for i in 1..final_number {
-            let difficulty = shared.calculate_difficulty(&parent).unwrap();
-            let mut uncles = vec![];
-            if i < 26 {
-                uncles.push(chain1[i as usize].clone().into());
-            }
-            let new_block = gen_block(&parent, i + 100, difficulty, vec![], uncles);
-            chain_controller
-                .process_block(Arc::new(new_block.clone()))
-                .expect("process block ok");
-            chain2.push(new_block.clone());
-            parent = new_block.header().clone();
-        }
-        let tip = shared.chain_state().read().tip_header().clone();
-        let total_uncles_count = shared.block_ext(&tip.hash()).unwrap().total_uncles_count;
-        assert_eq!(total_uncles_count, 25);
-        let difficulty = shared.calculate_difficulty(&tip).unwrap();
-
-        // 25 * 10 * 1000 / 200
-        assert_eq!(difficulty, U256::from(1250u64));
-
-        let (chain_controller, shared) = start_chain(Some(consensus.clone()));
-        let mut chain2: Vec<Block> = Vec::new();
-        for i in 1..final_number - 1 {
-            chain_controller
-                .process_block(Arc::new(chain1[(i - 1) as usize].clone()))
-                .expect("process block ok");
-        }
-
-        parent = shared.block_header(&shared.block_hash(0).unwrap()).unwrap();
-        for i in 1..final_number {
-            let difficulty = shared.calculate_difficulty(&parent).unwrap();
-            let mut uncles = vec![];
-            if i < 11 {
-                uncles.push(chain1[i as usize].clone().into());
-            }
-            let new_block = gen_block(&parent, i + 100, difficulty, vec![], uncles);
-            chain_controller
-                .process_block(Arc::new(new_block.clone()))
-                .expect("process block ok");
-            chain2.push(new_block.clone());
-            parent = new_block.header().clone();
-        }
-        let tip = shared.chain_state().read().tip_header().clone();
-        let total_uncles_count = shared.block_ext(&tip.hash()).unwrap().total_uncles_count;
-        assert_eq!(total_uncles_count, 10);
-        let difficulty = shared.calculate_difficulty(&tip).unwrap();
-
-        // min[10 * 10 * 1000 / 200, 1000]
-        assert_eq!(difficulty, U256::from(1000u64));
-
-        let (chain_controller, shared) = start_chain(Some(consensus.clone()));
-        let mut chain2: Vec<Block> = Vec::new();
-        for i in 1..final_number - 1 {
-            chain_controller
-                .process_block(Arc::new(chain1[(i - 1) as usize].clone()))
-                .expect("process block ok");
-        }
-
-        parent = shared.block_header(&shared.block_hash(0).unwrap()).unwrap();
-        for i in 1..final_number {
-            let difficulty = shared.calculate_difficulty(&parent).unwrap();
-            let mut uncles = vec![];
-            if i < 151 {
-                uncles.push(chain1[i as usize].clone().into());
-            }
-            let new_block = gen_block(&parent, i + 100, difficulty, vec![], uncles);
-            chain_controller
-                .process_block(Arc::new(new_block.clone()))
-                .expect("process block ok");
-            chain2.push(new_block.clone());
-            parent = new_block.header().clone();
-        }
-        let tip = shared.chain_state().read().tip_header().clone();
-        let total_uncles_count = shared.block_ext(&tip.hash()).unwrap().total_uncles_count;
-        assert_eq!(total_uncles_count, 150);
-        let difficulty = shared.calculate_difficulty(&tip).unwrap();
-
-        // max[150 * 10 * 1000 / 200, 2 * 1000]
-        assert_eq!(difficulty, U256::from(2000u64));
     }
 }
