+++ chain/src/lib.rs
@@ -9,3 +9,5 @@
 
 pub mod chain;
 pub mod error;
+#[cfg(test)]
+mod tests;
